cmd/microcloud/ask: Skip systems which aren't marked for local storage contribution
When growing a cluster (microcloud add), it might be that the already existing members don't anymore have any left over disks.
In this case they aren't marked to contribute disks to the local storage when adding new cluster members.

Therefore we can skip those and only ever exit in case a member which should contribute local storage doesn't have any disks.
That is the case when creating a new cluster or when adding a cluster member which cannot fulfill the requirement of the
already existing cluster.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>